server_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pymongo==4.9.1
motor==3.6.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.17
email-validator==2.2.0

# AI/ML Dependencies for RAG Pipeline
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.4
sentence-transformers==2.3.1
openai==2.15.0
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2

# OCR Dependencies for scanned/image PDFs
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
